+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 2021/05/13 v0.2.6
+
+* New functions:
+  
+  - FileVersionTable subclasses now can choose not to check the expected output files and use all produced
+    files in the versioned dir instead.
+  - Added FITS to raw version table
+  
+* Function improvements:
+  
+  - Improved HEVC and VVC compilation scripts
+  - Improved support for datasets consisting of symbolic links to a single copy of the dataset
+  - Improved default wrapper codec names when no hexdump signature is desired
+  
+* Plotting improvements:
+  
+  - Improved the general plot function to increase control over displayed colors.
+  - Added ScalarToScalarAnalyzer to plot dictionary cell data. Shadowed bands based on std can now be depicted.
+  - Several minor fix-ups to plot rendering and general stability
+
+* Docs are now displayed on the public site automatically point to the dev banch
+
 # 2021/04/27 v0.2.5
 
 - Improved enb compatibility with Windows and MacOS